requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Click==7.0
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
django-simple-history==3.3.0
django-webpack-loader==1.8.1
djangorestframework==3.14.0
fonttools==4.39.0
graphviz==0.20.1
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
nodejs==0.1.1
numpy==1.24.2
optional-django==0.1.0
packaging==23.0
Pillow==9.4.0
ply==3.11
progressbar2==4.2.0
psycopg2==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-utils==3.5.2
pytz==2022.7.1
PyYAML==6.0
quantiphy==2.19
react==4.3.0
requests==2.28.2
scipy==1.10.1
setuptools-scm==7.1.0
six==1.16.0
sqlparse==0.4.3
tabulate==0.9.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14